Use allOf workaround for additional ref properties
Both OpenAPI 2 and 3 (and JSON References, which they use) don't allow
properties to be added to $ref objects.  This makes it difficult to
define property-specific attributes for a referenced type (e.g.
property description, deprecated, xml not for all uses of type).
Use the [workaround] suggested by Ivan Goncharov to add attributes to an
allOf object with the $ref as its only member.

Note: This works with OpenAPI Generator, but does not work with Autorest
due to Azure/autorest#2652, which can be worked around by moving the
attributes to the $ref object (which violates OAS, but is supported by
Autorest) as a pre-build step.

[workaround]: https://github.com/OAI/OpenAPI-Specification/issues/556#issuecomment-192007034

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>