Improve report operations
Both standard and custom reports can return only a few media types
(based on the specified `reportFormat`).  Define the types explicitly as
documentation for users.  Include `text/csv`, even though it is not
produced, so it will be sent in `Accept` in hopes it will be used in the
future.

Since existing code generators don't handle responses with different
schemas for different content types well (usually all but one are
ignored), create separate operations with hard-coded `?format=JSON` and
a defined JSON schema (the same schema, `CustomReport` is returned for
both) so that generated clients can use the response from these
operations or call the `*InFormat` operation to get the report as a
binary stream of the requested format.

Note: Although neither OpenAPI 2 nor 3 forbid query parameters in path
strings (as far as I can see), it doesn't seem likely to be intended
and generators are not likely to support it.  A build step moves these
paths to [x-ms-paths] for Autorest, which supports it as an extension.

[x-ms-paths]: https://github.com/Azure/autorest/tree/master/docs/extensions#x-ms-paths

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>