Inline map-only schemas
A schema which has `additionalProperties` without `properties` generates
a map/dictionary type, so the schema name is not useful.

The build process must remove schemas from schemas that have
`additionalProperties` with `properties` to work around bugs in
Autorest, which is more difficult if `additionalProperties` can be a
`$ref` (since a named schema may or may not require replacement based on
whether any `$ref` are from `properties` - and may require splitting if
it has both types).  Inline the schemas to avoid the hassle.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>