Move path parameters to operations
The order of parameters in generated method signatures is not specified.
OpenAPI Generator puts path parameters before operation parameters,
Autorest does the opposite.

For consistency and so that the generated code has the most natural
parameter order for each particular operation, define all parameters on
the operation in the appropriate order.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>