Remove required parent properties for autorest2
Although OpenAPI allows properties of a parent schema to be required in
a child schema, Autorest does not enforce the requirement and OAV
produces OBJECT_MISSING_REQUIRED_PROPERTY_DEFINITION errors.

This occurs for EmployeePayStub which derives from EmployeeWithholding
and makes several properties required (e.g. fedWithholding).

NullableNotRequiredTransformer was run with requireUnconstrained: true
to preserve the requirements.  Remove it to avoid the errors.  Add a
note about the issue.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>